+++ src/main/java/com/arcadia/adminpanel/util/WarnManager.java
@@ -1,4 +1,4 @@
-package com.vyrriox.arcadiaadminpanel.util;
+package com.arcadia.adminpanel.util;
 
 import com.arcadia.lib.ServerContext;
 import com.arcadia.lib.data.DatabaseManager;